requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dnspython==2.0.0
fastapi==0.63.0
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
mongotransactions==1.2.1
pycparser==2.20
pydantic==1.8.1
pymongo==3.11.3
python-dotenv==0.15.0
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
